+++ core/libs/solvespace_translate.py
@@ -7,263 +7,277 @@ __copyright__ = "Copyright (C) 2016-2019"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
-from typing import Tuple, List, Union
-from math import radians, cos, sin
-from .pyslvs import VJoint, VPoint
-from .python_solvespace import (
-    # Entities & Constraint
-    Point3d,
-    Workplane,
-    Normal3d,
-    Point2d,
-    LineSegment2d,
-    Constraint,
-    # Result flags
-    SLVS_RESULT_OKAY,
-    SLVS_RESULT_INCONSISTENT,
-    SLVS_RESULT_DIDNT_CONVERGE,
-    # System base
-    System,
-    groupNum,
-    Slvs_MakeQuaternion,
+from typing import (
+    Tuple,
+    List,
+    Sequence,
+    Dict,
+    Union,
 )
+from math import radians, cos, sin
+from .pyslvs import (
+    get_vlinks,
+    VJoint,
+    VPoint,
+    VLink,
+)
+from .python_solvespace import ResultFlag, Entity, SolverSystem
 
 _Coord = Tuple[float, float]
 
 
-def _2d_system(num: int) -> Tuple[System, Workplane, LineSegment2d]:
-    """Create 2D CAD system."""
-    sys = System(num + 11)
-    sys.default_group = groupNum(1)
-    origin = Point3d(sys.add_param(0.), sys.add_param(0.), sys.add_param(0.))
-    qw, qx, qy, qz = Slvs_MakeQuaternion(1, 0, 0, 0, 1, 0)
-    wp1 = Workplane(origin, Normal3d(
-        sys.add_param(qw),
-        sys.add_param(qx),
-        sys.add_param(qy),
-        sys.add_param(qz)
-    ))
-    origin2d = Point2d(wp1, sys.add_param(0.), sys.add_param(0.))
-    Constraint.dragged(wp1, origin2d)
-    hp = Point2d(wp1, sys.add_param(10.), sys.add_param(0.))
-    Constraint.dragged(wp1, hp)
-    # Name 'ground' is a horizontal line through (0, 0) and (10, 0).
-    h_line = LineSegment2d(wp1, origin2d, hp)
-    sys.default_group = groupNum(2)
-    return sys, wp1, h_line
-
-
-def _pos(p: Point2d) -> _Coord:
-    """Get position of a Point2d instance."""
-    return p.u().value, p.v().value
-
-
 def slvs_solve(
-    vpoints: Tuple[VPoint],
-    inputs: Tuple[Tuple[int, int, float], ...]
+    vpoints: Sequence[VPoint],
+    inputs: Dict[Tuple[int, int], float]
 ) -> Tuple[List[Union[_Coord, Tuple[_Coord, _Coord]]], int]:
     """Use element module to convert into solvespace expression."""
     if not vpoints:
         return [], 0
 
-    # Define VLinks here.
-    vlinks = {}
+    vlinks: Dict[str, VLink] = {vlink.name: vlink for vlink in get_vlinks(vpoints)}
+
+    # Solvespace kernel
+    sys = SolverSystem()
+    sys.set_group(1)
+    wp = sys.create_2d_base()
+    origin_2d = sys.add_point_2d(0., 0., wp)
+    sys.dragged(origin_2d, wp)
+    origin_2d_p = sys.add_point_2d(10., 0., wp)
+    sys.dragged(origin_2d_p, wp)
+    hv = sys.add_line_2d(origin_2d, origin_2d_p, wp)
+    sys.set_group(2)
+
+    points: List[Entity] = []
+    sliders: Dict[int, int] = {}
+    slider_bases: List[Entity] = []
+    slider_slots: List[Entity] = []
+
     for i, vpoint in enumerate(vpoints):
-        for vlink in vpoint.links:
-            if vlink == 'ground':
-                continue
-            if vlink in vlinks:
-                if i not in vlinks[vlink]:
-                    vlinks[vlink].append(i)
-            else:
-                vlinks[vlink] = [i]
-
-    # Limitation of Solvespace kernel sys.
-    point_count = 0
-    for vpoint in vpoints:
-        if vpoint.type == VJoint.R:
-            point_count += 1
-        elif vpoint.type in {VJoint.P, VJoint.RP}:
-            point_count += 3
-
-    sys, wp1, h_line = _2d_system(point_count * 2 + len(inputs) * 2)
-
-    solved_points = []
-    slider_points = {}
-    slot_points = {}
-    for i, vpoint in enumerate(vpoints):
-        if vpoint.type == VJoint.R:
-            # Point coordinate.
-            x = sys.add_param(vpoint.cx)
-            y = sys.add_param(vpoint.cy)
-            solved_points.append(Point2d(wp1, x, y))
-            if vpoint.grounded():
-                Constraint.dragged(wp1, solved_points[i])
-        elif vpoint.type in {VJoint.P, VJoint.RP}:
-            # Base point coordinate.
-            bx = sys.add_param(vpoint.c[0][0])
-            by = sys.add_param(vpoint.c[0][1])
-            slider_points[i] = Point2d(wp1, bx, by)
-            # Slot coordinate.
-            angle = radians(vpoint.angle)
-            sx = sys.add_param(vpoint.c[0][0] + cos(angle))
-            sy = sys.add_param(vpoint.c[0][1] + sin(angle))
-            slot_points[i] = Point2d(wp1, sx, sy)
-            # Pin coordinate.
-            px = vpoint.c[1][0]
-            py = vpoint.c[1][1]
-            if vpoint.has_offset() and vpoint.true_offset() == 0:
-                value = 0.1 if vpoint.true_offset() > 0 else -0.1
-                px += value
-                py += value
-            x = sys.add_param(px)
-            y = sys.add_param(py)
-            solved_points.append(Point2d(wp1, x, y))
-            if vpoint.grounded():
-                Constraint.dragged(wp1, slider_points[i])
-            elif vpoint.pin_grounded():
-                Constraint.dragged(wp1, solved_points[i])
-
-    # Link constraints.
-    for vlink in vlinks:
-        if len(vlinks[vlink]) < 2:
+        if vpoint.no_link():
+            x, y = vpoint.c[0]  # type: float
+            point = sys.add_point_2d(x, y, wp)
+            sys.dragged(point, wp)
+            points.append(point)
             continue
-        a = vlinks[vlink][0]
-        b = vlinks[vlink][1]
-        distance = vpoints[a].distance(vpoints[b])
-        if vpoints[a].is_slot_link(vlink):
-            p1 = slider_points[a]
-        else:
-            p1 = solved_points[a]
-        if vpoints[b].is_slot_link(vlink):
-            p2 = slider_points[b]
-        else:
-            p2 = solved_points[b]
-        if distance:
-            Constraint.distance(distance, wp1, p1, p2)
-        else:
-            Constraint.on(wp1, p1, p2)
-        for c in vlinks[vlink][2:]:
-            for d in (a, b):
-                distance = vpoints[c].distance(vpoints[d])
-                if vpoints[c].is_slot_link(vlink):
-                    p1 = slider_points[c]
-                else:
-                    p1 = solved_points[c]
-                if vpoints[d].is_slot_link(vlink):
-                    p2 = slider_points[d]
-                else:
-                    p2 = solved_points[d]
-                if distance:
-                    Constraint.distance(distance, wp1, p1, p2)
-                else:
-                    Constraint.on(wp1, p1, p2)
 
-    # Slider constraints.
-    for i in slider_points:
-        vpoint = vpoints[i]
-        slider_slot = LineSegment2d(wp1, slider_points[i], slot_points[i])
         if vpoint.grounded():
-            # Slot is grounded.
-            if vpoint.angle in {0., 180.}:
-                Constraint.parallel(wp1, h_line, slider_slot)
+            x, y = vpoint.c[0]
+            if vpoint.type in {VJoint.P, VJoint.RP}:
+                sliders[i] = len(slider_bases)
+                # Base point (slot) is fixed.
+                point = sys.add_point_2d(x, y, wp)
+                sys.dragged(point, wp)
+                slider_bases.append(point)
+                # Slot point (slot) is movable.
+                x += cos(vpoint.angle)
+                y += sin(vpoint.angle)
+                slider_slots.append(sys.add_point_2d(x, y, wp))
+                # Pin is movable.
+                x, y = vpoint.c[1]
+                if vpoint.has_offset() and vpoint.true_offset() <= 0.1:
+                    if vpoint.offset() > 0:
+                        x += 0.1
+                        y += 0.1
+                    else:
+                        x -= 0.1
+                        y -= 0.1
+                points.append(sys.add_point_2d(x, y, wp))
             else:
-                Constraint.angle(wp1, vpoint.angle, h_line, slider_slot)
-            Constraint.on(wp1, solved_points[i], slider_slot)
-            if vpoint.has_offset():
-                offset = vpoint.offset()
-                if vpoint.offset() > 0:
-                    Constraint.distance(offset, wp1, slider_points[i], solved_points[i])
+                point = sys.add_point_2d(x, y, wp)
+                sys.dragged(point, wp)
+                points.append(point)
+            continue
+
+        x, y = vpoint.c[0]
+        point = sys.add_point_2d(x, y, wp)
+        if vpoint.type in {VJoint.P, VJoint.RP}:
+            sliders[i] = len(slider_bases)
+            # Base point (slot) is movable.
+            slider_bases.append(point)
+            # Slot point (slot) is movable.
+            x += cos(vpoint.angle)
+            y += sin(vpoint.angle)
+            slider_slots.append(sys.add_point_2d(x, y, wp))
+            if vpoint.pin_grounded():
+                # Pin is fixed.
+                x, y = vpoint.c[1]
+                point = sys.add_point_2d(x, y, wp)
+                sys.dragged(point, wp)
+                points.append(point)
+            else:
+                # Pin is movable.
+                x, y = vpoint.c[1]
+                if vpoint.has_offset() and vpoint.true_offset() <= 0.1:
+                    if vpoint.offset() > 0:
+                        x += 0.1
+                        y += 0.1
+                    else:
+                        x -= 0.1
+                        y -= 0.1
+                points.append(sys.add_point_2d(x, y, wp))
+            continue
+
+        # Point is movable.
+        points.append(point)
+
+    for vlink in vlinks.values():
+        if len(vlink.points) < 2:
+            continue
+        if vlink.name == 'ground':
+            continue
+
+        a = vlink.points[0]
+        b = vlink.points[1]
+        vp1 = vpoints[a]
+        vp2 = vpoints[b]
+        if vp1.is_slot_link(vlink.name):
+            p1 = slider_bases[sliders[a]]
+        else:
+            p1 = points[a]
+        if vp2.is_slot_link(vlink.name):
+            p2 = slider_bases[sliders[b]]
+        else:
+            p2 = points[b]
+        sys.distance(p1, p2, vp1.distance(vp2), wp)
+
+        for c in vlink.points[2:]:  # type: int
+            for d in (a, b):
+                vp1 = vpoints[c]
+                vp2 = vpoints[d]
+                if vp1.is_slot_link(vlink.name):
+                    p1 = slider_bases[sliders[c]]
                 else:
-                    Constraint.distance(offset, wp1, solved_points[i], slider_points[i])
+                    p1 = points[c]
+                if vp2.is_slot_link(vlink.name):
+                    p2 = slider_bases[sliders[d]]
+                else:
+                    p2 = points[d]
+                sys.distance(p1, p2, vp1.distance(vp2), wp)
+
+    for a, b in sliders.items():
+        # Base point
+        vp1 = vpoints[a]
+        p1 = points[a]
+        # Base slot
+        slider_slot = sys.add_line_2d(slider_bases[b], slider_slots[b], wp)
+        if vp1.grounded():
+            # Slot is grounded.
+            sys.angle(hv, slider_slot, vp1.angle, wp)
+            sys.coincident(p1, slider_slot, wp)
+            if vp1.has_offset():
+                p2 = slider_bases[b]
+                if vp1.offset():
+                    sys.distance(p2, p1, vp1.offset(), wp)
+                else:
+                    sys.coincident(p2, p1, wp)
         else:
             # Slider between links.
-            for vlink in vpoint.links[:1]:
-                f1 = vlinks[vlink][0]
-                if f1 == i:
-                    if len(vlinks[vlink]) < 2:
+            for name in vp1.links[:1]:  # type: str
+                vlink = vlinks[name]
+                # A base link friend.
+                c = vlink.points[0]
+                if c == a:
+                    if len(vlink.points) < 2:
                         # If no any friend.
                         continue
-                    f1 = vlinks[vlink][1]
-                if vpoints[f1].is_slot_link(vlink):
-                    helper = LineSegment2d(wp1, slider_points[i], slider_points[f1])
+                    c = vlink.points[1]
+
+                vp2 = vpoints[c]
+                if vp2.is_slot_link(vlink.name):
+                    # c is a slider, and it is be connected with slot link.
+                    p2 = slider_bases[sliders[c]]
                 else:
-                    helper = LineSegment2d(wp1, slider_points[i], solved_points[f1])
-                angle = vpoint.slope_angle(vpoints[f1]) - vpoint.angle
-                if angle in {0., 180.}:
-                    Constraint.parallel(wp1, slider_slot, helper)
-                else:
-                    Constraint.angle(wp1, angle, slider_slot, helper)
-                Constraint.on(wp1, solved_points[i], slider_slot)
-                if vpoint.has_offset():
-                    offset = vpoint.offset()
-                    if vpoint.offset() > 0:
-                        Constraint.distance(offset, wp1, slider_points[i], solved_points[i])
+                    # c is a R joint or it is not connected with slot link.
+                    p2 = points[c]
+                sys.angle(
+                    slider_slot,
+                    sys.add_line_2d(slider_bases[b], p2, wp),
+                    vp1.slope_angle(vp2) - vp1.angle,
+                    wp
+                )
+                sys.coincident(p1, slider_slot, wp)
+
+                if vp1.has_offset():
+                    p2 = slider_bases[b]
+                    if vp1.offset():
+                        sys.distance(p2, p1, vp1.offset(), wp)
                     else:
-                        Constraint.distance(offset, wp1, solved_points[i], slider_points[i])
+                        sys.coincident(p2, p1, wp)
 
-        if vpoint.type == VJoint.P:
-            for vlink in vpoint.links[1:]:
-                f1 = vlinks[vlink][0]
-                if f1 == i:
-                    if len(vlinks[vlink]) < 2:
+            if vp1.type != VJoint.P:
+                continue
+
+            for name in vp1.links[1:]:
+                vlink = vlinks[name]
+                # A base link friend.
+                c = vlink.points[0]
+                if c == a:
+                    if len(vlink.points) < 2:
                         # If no any friend.
                         continue
-                    f1 = vlinks[vlink][1]
-                if vpoints[f1].is_slot_link(vlink):
-                    helper = LineSegment2d(wp1, solved_points[i], slider_points[f1])
-                else:
-                    helper = LineSegment2d(wp1, solved_points[i], solved_points[f1])
-                angle = vpoint.slope_angle(vpoints[f1]) - vpoint.angle
-                if angle in {0., 180.}:
-                    Constraint.parallel(wp1, slider_slot, helper)
-                else:
-                    Constraint.angle(wp1, angle, slider_slot, helper)
+                    c = vlink.points[1]
 
-    for p0, p1, angle in inputs:
-        """The constraints of drive shaft.
-        
-        Simulate the input variables to the mechanism.
-        The 'base points' are shaft center.
-        """
-        if p0 == p1:
+                vp2 = vpoints[c]
+                if vp2.is_slot_link(vlink.name):
+                    # c is a slider, and it is be connected with slot link.
+                    p2 = slider_bases[sliders[c]]
+                else:
+                    # c is a R joint or it is not connected with slot link.
+                    p2 = points[c]
+                sys.angle(
+                    slider_slot,
+                    sys.add_line_2d(p1, p2, wp),
+                    vp1.slope_angle(vp2) - vp1.angle,
+                    wp
+                )
+
+    for (b, d), angle in inputs.items():
+        # The constraints of drive shaft.
+        # Simulate the input variables to the mechanism.
+        # The 'base points' are shaft center.
+        if b == d:
             continue
 
-        if vpoints[p0].type == VJoint.R:
-            p_base = solved_points[p0]
+        vp1 = vpoints[b]
+        if vp1.type == VJoint.R:
+            p1 = points[b]
         else:
-            p_base = slider_points[p0]
+            p1 = slider_bases[sliders[b]]
 
         angle = radians(angle)
-        x = sys.add_param(vpoints[p0].cx + cos(angle))
-        y = sys.add_param(vpoints[p0].cy + sin(angle))
-        p_hand = Point2d(wp1, x, y)
-        Constraint.dragged(wp1, p_hand)
+        p2 = sys.add_point_2d(vp1.cx + cos(angle), vp1.cy + sin(angle), wp)
+        sys.dragged(p2, wp)
         # The virtual link that dragged by "hand".
-        leader = LineSegment2d(wp1, p_base, p_hand)
+        leader = sys.add_line_2d(p1, p2, wp)
         # Make another virtual link that should follow "hand".
-        if vpoints[p1].type == VJoint.R:
-            p_drive = solved_points[p1]
+        vp2 = vpoints[d]
+        if vp2.type == VJoint.R:
+            p2 = points[d]
         else:
-            p_drive = slider_points[p1]
-        link = LineSegment2d(wp1, p_base, p_drive)
-        Constraint.angle(wp1, 0.5, link, leader)
+            p2 = slider_bases[sliders[d]]
+        link = sys.add_line_2d(p1, p2, wp)
+        sys.angle(link, leader, 0.5, wp)
 
     # Solve
     result_flag = sys.solve()
-    if result_flag == SLVS_RESULT_OKAY:
+    if result_flag == ResultFlag.OKAY:
         result_list = []
         for i, vpoint in enumerate(vpoints):
-            p = solved_points[i]
+            p1 = points[i]
+            x1, y1 = sys.params(p1.params)
             if vpoint.type == VJoint.R:
-                result_list.append(_pos(p))
+                result_list.append((x1, y1))
             else:
-                result_list.append((_pos(slider_points[i]), _pos(p)))
-        return result_list, sys.dof
-    elif result_flag == SLVS_RESULT_INCONSISTENT:
-        error = "Inconsistent."
-    elif result_flag == SLVS_RESULT_DIDNT_CONVERGE:
-        error = "Did not converge."
+                p2 = slider_bases[sliders[i]]
+                x2, y2 = sys.params(p2.params)
+                result_list.append(((x2, y2), (x1, y1)))
+        return result_list, sys.dof()
+
+    if result_flag == ResultFlag.INCONSISTENT:
+        error = "inconsistent"
+    elif result_flag == ResultFlag.DIDNT_CONVERGE:
+        error = "didn't converge"
     else:
-        error = "Too many unknowns."
+        error = "too many unknowns"
+    error += f": {sys.faileds()}\n{sys.constraints()}"
     raise ValueError(error)